chore: Add trending movie list and image URL constant
This commit adds the trendingMovieList constant to the constant.js file, which stores the URL for retrieving the trending movies from the API. Additionally, it adds the imageMovie constant, which stores the base URL for retrieving movie images. These changes are necessary to support the display of trending movies and their corresponding images in the application.