feat: Update npm dependencies and add react-icons package
This commit updates the npm dependencies in the package.json file to include the latest stable versions of react, react-dom, and react-helmet. Additionally, it adds the react-icons package with version 5.2.1. These changes are necessary to ensure compatibility with the latest features and improvements in the dependencies and to introduce the use of react-icons in the application.